+++ extensions/chocolatey-core.extension/extensions/Get-AvailableDriveLetter.ps1
@@ -22,6 +22,7 @@
   http://stackoverflow.com/questions/12488030/getting-a-free-drive-letter/29373301#29373301
 #>
 function Get-AvailableDriveLetter {
+  [CmdletBinding()]
   param (
     [char[]]$ExcludedLetters,